AnimatedStack becomes HomeScreen, code cleanup
#30
- Deleted AnimatedStack.tsx component, and moved all of its code into HomeScreen. (Previously, HomeScreen was just returning an AnimatedStack anyway, so we merged this into one file called HomeScreen).
- Deleted unused constants that were previously used for card animations
- Deleted useCardAnimation and usePanGesture since these are now depreciated since we are now using the Swiper component.
